+++ templates/types/streaming/fastapi/app/api/routers/upload.py
@@ -1,5 +1,5 @@
 import logging
-from typing import List
+from typing import List, Any
 
 from fastapi import APIRouter, HTTPException
 from pydantic import BaseModel
@@ -13,13 +13,17 @@ logger = logging.getLogger("uvicorn")
 
 class FileUploadRequest(BaseModel):
     base64: str
+    filename: str
+    params: Any = None
 
 
 @r.post("")
 def upload_file(request: FileUploadRequest) -> List[str]:
     try:
         logger.info("Processing file")
-        return PrivateFileService.process_file(request.base64)
+        return PrivateFileService.process_file(
+            request.filename, request.base64, request.params
+        )
     except Exception as e:
         logger.error(f"Error processing file: {e}", exc_info=True)
         raise HTTPException(status_code=500, detail="Error processing file")